feat: adopt LazyMatcher, LazyString, & StringObject
StringObject is used to support comparisions for String. The main
purpose is to make 'writing automation scripts' easily and friendly.

LazyMatcher is used to workaround the problem that objects retrieved
from UtaPassUtil.withIndex() cannot be reused. Thus, evaluation is
conducted at the last monent.

LazyString is based on LazyMatcher and contains StringObject methods.